requirements.txt
openai>=1.0.0
Pillow>=11.0.0
numpy>=1.24.3
scikit-learn>=1.3.2
requests>=2.31.0
streamlit>=1.28.0
python-dotenv>=1.0.0
